// src/components/Toast/types.tsx
export type ToastShowParams = ToastData & ToastOptions;

export type ToastHideParams = void;

export type NotificationToastProps = BaseToastProps & {
  time?:number | object
  timeStyle?:ViewStyle
};

export type BaseToastProps = {
  title?: string;
  message?: string;
  onPress?: () => void;
  activeOpacity?: number;
  style?: ViewStyle;
  touchableContainerProps?: TouchableOpacityProps;
  contentContainerStyle?: ViewStyle;
  contentContainerProps?: ViewStyle;
  titleStyle?: TextStyle;
  titleNumberOfLines?: number;
  titleProps?: TextStyle;
  messageStyle?: TextStyle;
  messageNumberOfLines?: number;
  messageProps?: TextStyle;
  renderLeadingIcon?: () => React.ReactNode;
  renderTrailingIcon?: () => React.ReactNode;
  show: (params: ToastShowParams) => void;
  hide: (params: ToastHideParams) => void;
};